refactor: simplify event
