+++ nexus-common/src/models/event/mod.rs
@@ -1,185 +1,30 @@
 mod errors;
 
-use crate::{
-    db::{kv::RedisResult, RedisOps},
-    universal_tag::homeserver_parsed_uri::HomeserverParsedUri,
-};
-use pubky::Event as StreamEvent;
-use pubky_app_specs::Resource;
+use crate::db::{kv::RedisResult, RedisOps};
 use serde::{Deserialize, Serialize};
-use std::{fmt, path::PathBuf};
-use tracing::{debug, error, warn};
+use tracing::error;
 
 pub use errors::EventProcessorError;
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum EventType {
-    Put,
-    Del,
-}
-
-impl From<pubky::EventType> for EventType {
-    fn from(value: pubky::EventType) -> Self {
-        match value {
-            pubky::EventType::Put { .. } => Self::Put,
-            pubky::EventType::Delete => Self::Del,
-        }
-    }
-}
-
-/// Result of parsing an event line from a homeserver.
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug)]
-pub enum ParseResult {
-    /// Successfully parsed into a known, actionable event.
-    Parsed(Event),
-    /// Known resource type that Nexus does not handle (e.g. LastRead, Feed, Blob).
-    Skipped,
-    /// URI was not recognised by pubky-app-specs. This may be an app-specific
-    /// path (e.g. `/pub/mapky/tags/...`) or a genuinely malformed URI.
-    /// Callers should attempt fallback handling and log `reason` if no handler claims it.
-    UnrecognizedUri {
-        event_type: EventType,
-        uri: String,
-        reason: String,
-    },
-}
-
-impl ParseResult {
-    fn unrecognized_uri(event_type: EventType, uri: String, reason: String) -> Self {
-        Self::UnrecognizedUri {
-            event_type,
-            uri,
-            reason,
-        }
-    }
-}
-
-impl fmt::Display for EventType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let upper_case_str = match self {
-            EventType::Put => "PUT",
-            EventType::Del => "DEL",
-        };
-        write!(f, "{upper_case_str}")
-    }
-}
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Event {
-    /// Pubky resource URI from the homeserver event line.
-    pub uri: String,
+pub struct RawEvent(pub String);
 
-    /// Operation represented by the event, used to dispatch to PUT or DEL handlers.
-    pub event_type: EventType,
-
-    /// Parsed representation of [`Self::uri`].
-    pub parsed_uri: HomeserverParsedUri,
-
-    /// Local files directory on Nexus used for file-backed events.
-    pub files_path: PathBuf,
-
-    /// Original event line as received from the homeserver.
-    event_line: String,
+#[async_trait::async_trait]
+impl RedisOps for RawEvent {
+    async fn prefix() -> String {
+        "Event".to_string()
+    }
 }
 
-impl RedisOps for Event {}
-
-impl AsRef<[String]> for Event {
+impl AsRef<[String]> for RawEvent {
     fn as_ref(&self) -> &[String] {
-        std::slice::from_ref(&self.event_line)
+        std::slice::from_ref(&self.0)
     }
 }
 
-impl Event {
-    /// Parse event from a line returned by the homeserver's `/events` endpoint.
-    pub fn parse_event(
-        line: &str,
-        files_path: PathBuf,
-    ) -> Result<ParseResult, EventProcessorError> {
-        debug!("New event: {}", line);
-        let parts: Vec<&str> = line.split(' ').collect();
-        if parts.len() != 2 {
-            return Err(EventProcessorError::InvalidEventLine(format!(
-                "Malformed event line, {line}"
-            )));
-        }
-
-        let event_type = match parts[0] {
-            "PUT" => Ok(EventType::Put),
-            "DEL" => Ok(EventType::Del),
-            other => Err(EventProcessorError::InvalidEventLine(format!(
-                "Unknown event type: {other}"
-            ))),
-        }?;
-
-        let uri = parts[1].to_string();
-        let event_line = line.to_string();
-
-        Self::parse_event_parts(event_type, uri, event_line, files_path)
-    }
-
-    /// Constructs a nexus [`Event`] directly from a [`StreamEvent`], avoiding
-    /// the string round-trip through [`Self::parse_event`].
-    pub fn from_stream_event(
-        stream_event: &StreamEvent,
-        files_path: PathBuf,
-    ) -> Result<Option<Self>, EventProcessorError> {
-        let event_type: EventType = stream_event.event_type.clone().into();
-
-        let uri = stream_event.resource.to_pubky_url();
-        debug!("New stream event: {event_type} {uri}");
-
-        let event_line = format!("{event_type} {uri}");
-        match Self::parse_event_parts(event_type, uri, event_line, files_path)? {
-            ParseResult::Parsed(event) => Ok(Some(event)),
-            ParseResult::Skipped => Ok(None),
-            ParseResult::UnrecognizedUri { reason, .. } => {
-                warn!("Unrecognized event URI: {reason}");
-                Ok(None)
-            }
-        }
-    }
-
-    fn parse_event_parts(
-        event_type: EventType,
-        uri: String,
-        event_line: String,
-        files_path: PathBuf,
-    ) -> Result<ParseResult, EventProcessorError> {
-        // Validate and parse the URI using HomeserverParsedUri. This handles both
-        // standard pubky-app-specs URIs and universal tag URIs from other apps.
-        let parsed_uri = match HomeserverParsedUri::try_from(uri.as_str()) {
-            Ok(parsed) => parsed,
-            Err(e) => return Ok(ParseResult::unrecognized_uri(event_type, uri, e)),
-        };
-
-        if let HomeserverParsedUri::AppSpec { resource, .. } = &parsed_uri {
-            match resource {
-                Resource::Unknown => {
-                    return Err(EventProcessorError::InvalidEventLine(format!(
-                        "Unknown resource in URI: {uri}"
-                    )))
-                }
-                Resource::LastRead | Resource::Feed(_) | Resource::Blob(_) => {
-                    return Ok(ParseResult::Skipped)
-                }
-                _ => (),
-            }
-        }
-
-        Ok(ParseResult::Parsed(Event {
-            uri,
-            event_type,
-            parsed_uri,
-            files_path,
-            event_line,
-        }))
-    }
-
-    /// Stores event line in Redis as part of the events list.
+impl RawEvent {
     #[tracing::instrument(name = "event.index.write", skip_all)]
-    pub async fn store_event(&self) -> RedisResult<()> {
+    pub async fn store(&self) -> RedisResult<()> {
         self.put_index_list(&["Events"]).await
     }
 
@@ -188,10 +33,8 @@ impl Event {
         limit: usize,
     ) -> RedisResult<(Vec<String>, u64)> {
         let start = cursor.unwrap_or(0);
-        // Clamp to usize::MAX: on 32-bit targets u64 can exceed usize; the LRANGE
-        // would return empty results for such a large index either way.
         let start_u = usize::try_from(start).unwrap_or(usize::MAX);
-        let result = Event::try_from_index_list(&["Events"], Some(start_u), Some(limit)).await;
+        let result = RawEvent::try_from_index_list(&["Events"], Some(start_u), Some(limit)).await;
 
         let events = match result {
             Ok(r) => r.unwrap_or_default(),
@@ -202,7 +45,6 @@ impl Event {
         };
 
         let next_cursor = start + events.len() as u64;
-
         Ok((events, next_cursor))
     }
 }
