+++ nexus-watcher/src/service/indexer/key_based.rs
@@ -1,8 +1,9 @@
 use std::{path::PathBuf, sync::Arc, time::Duration};
 
+use crate::events::Event;
 use futures::StreamExt;
 use nexus_common::db::PubkyConnector;
-use nexus_common::models::event::{Event, EventProcessorError};
+use nexus_common::models::event::EventProcessorError;
 use nexus_common::models::homeserver::Homeserver;
 use nexus_common::models::user::UserHsCursor;
 use pubky::{Event as StreamEvent, EventCursor, PublicKey};
