+++ nexus-watcher/tests/event_processor/bookmarks/raw.rs
@@ -4,7 +4,7 @@ use crate::event_processor::{
     users::utils::find_user_counts, utils::watcher::HomeserverHashIdPath,
 };
 use anyhow::Result;
-use nexus_common::models::event::Event;
+use nexus_common::models::event::RawEvent;
 use nexus_common::models::post::{Bookmark, PostStream};
 use pubky::Keypair;
 use pubky_app_specs::{
@@ -25,7 +25,7 @@ async fn test_homeserver_bookmark() -> Result<()> {
         status: None,
     };
     let user_id = test.create_user(&user_kp, &user).await?;
-    let (_, events_in_redis_before) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_before) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
 
     // Step 2: Create a post under that user
     let post = PubkyAppPost {
@@ -60,7 +60,7 @@ async fn test_homeserver_bookmark() -> Result<()> {
     assert_eq!(user_counts.bookmarks, 1);
 
     // INDEX_OP: Assert if the event writes the indexes
-    let (_, events_in_redis_after) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_after) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     assert!(events_in_redis_after > events_in_redis_before);
 
     let result_bookmarks = PostStream::get_bookmarked_posts(
