+++ nexus-watcher/tests/event_processor/files/create.rs
@@ -1,7 +1,7 @@
 use crate::event_processor::utils::watcher::{assert_file_details, WatcherTest};
 use anyhow::Result;
 use chrono::Utc;
-use nexus_common::models::event::Event;
+use nexus_common::models::event::RawEvent;
 use pubky::Keypair;
 use pubky_app_specs::traits::{HasIdPath, HashId};
 use pubky_app_specs::{blob_uri_builder, PubkyAppBlob, PubkyAppFile, PubkyAppUser};
@@ -28,7 +28,7 @@ async fn test_put_pubkyapp_file() -> Result<()> {
     let blob_relative_url = PubkyAppBlob::create_path(&blob_id);
     let blob_absolute_url = blob_uri_builder(user_id.clone(), blob_id);
 
-    let (_, events_in_redis_before) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_before) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
 
     test.create_file_from_body(&user_kp, blob_relative_url.as_str(), blob.0.clone())
         .await?;
@@ -52,7 +52,7 @@ async fn test_put_pubkyapp_file() -> Result<()> {
         blob_static_path.exists(),
         "File have to exist after PUT event"
     );
-    let (_, events_in_redis_after) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_after) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     assert!(events_in_redis_after > events_in_redis_before);
 
     Ok(())
