+++ nexus-watcher/tests/event_processor/files/delete.rs
@@ -1,7 +1,7 @@
 use crate::event_processor::utils::watcher::{assert_file_details, WatcherTest};
 use anyhow::Result;
 use chrono::Utc;
-use nexus_common::models::event::Event;
+use nexus_common::models::event::RawEvent;
 use nexus_common::models::{file::FileDetails, traits::Collection};
 use nexus_watcher::events::handlers;
 use pubky::Keypair;
@@ -33,7 +33,7 @@ async fn test_delete_pubkyapp_file() -> Result<()> {
     let blob_relative_url = PubkyAppBlob::create_path(&blob_id);
     let blob_absolute_url = blob_uri_builder(user_id.clone(), blob_id);
 
-    let (_, events_in_redis_before) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_before) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
 
     test.create_file_from_body(&user_kp, blob_relative_url.as_str(), blob.0.clone())
         .await?;
@@ -64,7 +64,7 @@ async fn test_delete_pubkyapp_file() -> Result<()> {
     let file_before_delete = files_before_delete[0].as_ref();
     assert!(file_before_delete.is_some());
 
-    let (_, events_in_redis_after) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_after) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     assert!(events_in_redis_after > events_in_redis_before);
 
     test.cleanup_file(&user_kp, &file_path).await?;
