+++ nexus-watcher/tests/event_processor/tags/post_put.rs
@@ -7,7 +7,7 @@ use crate::event_processor::users::utils::find_user_counts;
 use crate::event_processor::utils::watcher::{HomeserverHashIdPath, WatcherTest};
 use anyhow::Result;
 use chrono::Utc;
-use nexus_common::models::event::Event;
+use nexus_common::models::event::RawEvent;
 use nexus_common::models::notification::Notification;
 use nexus_common::models::post::PostDetails;
 use nexus_common::models::tag::post::TagPost;
@@ -54,7 +54,7 @@ async fn test_homeserver_put_tag_post() -> Result<()> {
     };
     let tag_path = tag.hs_path();
 
-    let (_, events_in_redis_before) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_before) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     // Put tag
     test.put(&user_kp, &tag_path, tag).await?;
 
@@ -71,7 +71,7 @@ async fn test_homeserver_put_tag_post() -> Result<()> {
     assert_eq!(post_tag.taggers[0], tagger_user_id);
 
     // CACHE_OP: Check if the tag is correctly cached
-    let (_, events_in_redis_after) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_after) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     assert!(events_in_redis_after > events_in_redis_before);
     let cache_post_tag = TagPost::get_from_index(
         &tagger_user_id,
