+++ nexus-watcher/tests/event_processor/users/raw.rs
@@ -5,7 +5,7 @@ use crate::{
     event_processor::utils::watcher::WatcherTest,
 };
 use anyhow::Result;
-use nexus_common::models::event::Event;
+use nexus_common::models::event::RawEvent;
 use nexus_common::{
     db::RedisOps,
     models::user::{UserCounts, UserSearch, USER_NAME_KEY_PARTS},
@@ -32,7 +32,7 @@ async fn test_homeserver_user_put_event() -> Result<()> {
         name: "Watcher:UserEvent:User".to_string(),
         status: Some("Running Nexus Watcher".to_string()),
     };
-    let (_, events_in_redis_before) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_before) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
 
     let user_id = test.create_user(&user_kp, &user).await?;
 
@@ -61,7 +61,7 @@ async fn test_homeserver_user_put_event() -> Result<()> {
 
     // CACHE_OP: Check if the event writes in the graph
     // User:Counts:user_id
-    let (_, events_in_redis_after) = Event::get_events_from_redis(None, 1000).await.unwrap();
+    let (_, events_in_redis_after) = RawEvent::get_events_from_redis(None, 1000).await.unwrap();
     assert!(events_in_redis_after > events_in_redis_before);
 
     let user_counts = UserCounts::get_from_index(&user_id)
