+++ nexus-watcher/tests/service/key_based_event_processor.rs
@@ -6,11 +6,12 @@ use std::time::Duration;
 use anyhow::Result;
 use chrono::Utc;
 use nexus_common::db::{exec_single_row, graph::Query, queries, PubkyClientError, RedisOps};
-use nexus_common::models::event::{Event, EventProcessorError};
+use nexus_common::models::event::EventProcessorError;
 use nexus_common::models::homeserver::Homeserver;
 use nexus_common::models::user::{user_hs_cursor_key, UserDetails};
 use nexus_common::types::DynError;
 use nexus_watcher::events::retry::{InitialBackoff, RetryScheduler};
+use nexus_watcher::events::Event;
 use nexus_watcher::events::EventHandler;
 use nexus_watcher::service::indexer::{KeyBasedEventProcessor, RunError, TEventProcessor};
 use nexus_watcher::service::runner::UserNotFoundBackoff;
