+++ nexus-watcher/tests/service/retry_processor.rs
@@ -4,12 +4,13 @@ use anyhow::Result;
 use chrono::Utc;
 use nexus_common::config::EventRetryConfig;
 use nexus_common::db::kv::RedisOps;
-use nexus_common::models::event::{EventProcessorError, EventType};
+use nexus_common::models::event::EventProcessorError;
 use nexus_watcher::events::retry::{
     RedisRetryStore, RetryEvent, RetryProcessor, RetryStore, RETRY_MANAGER_EVENTS_INDEX,
     RETRY_MANAGER_PREFIX,
 };
 use nexus_watcher::events::EventHandler;
+use nexus_watcher::events::EventType;
 use nexus_watcher::service::TEventProcessor;
 use pubky_app_specs::post_uri_builder;
 use std::path::PathBuf;
